pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "viflo"
version = "0.1.0-alpha.1"
dependencies = [
    "dnois",
    "lightning",
    "matplotlib",
    "numpy",
    "pyyaml",
    "tensorboard",
    "torch",
    "torchmetrics",
    "torchvision",
]
authors = [
    { name = "Jiaqi Guo", email = "gjq@tongji.edu.cn" }
]
description = "End-to-end design of computational cameras"
readme = "README.md"
license = "MIT"
scripts = { viflo = "viflo.run:main" }

[tool.setuptools]
packages = ["viflo"]